Create FAPAR
This code is for the processing of the S3-TOA-GPR-1.0 vegetation maps in GEE. This script allows you to visualize and investigate time series of different locations of your interest. You can copy the code in your own GEE code editor or you can run the link below:

https://code.earthengine.google.com/3225e78586a6444026e168b7b68bccf8

The code features the non-reconstructed S3-TOA-GPR-1.0 maps, the Whittaker smoother temporal reconstruction and the Gaussian Process Regression's Uncertainties too. The code is written in a way that you can change the temporal domain of your interest. Keep in mind that the visualization of the data on the global map is the mean pixel value of the images within your set temporal domain.